# docs/swagger.yaml
definitions:
...
  handlers.InviteCreatedResponse:
    properties:
      inviteToken:
        type: string
      url:
        type: string
    type: object
  handlers.LoginRequest:
    properties:
      altcha:
        example: eyJhbGdvcml0aG0iOi...
        type: string
      password:
        example: s3cr3t-p4ss
        type: string
      username:
        example: admin
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      token:
        example: eyJhbG...
        type: string
    type: object
...
  server.MemberInfo:
    properties:
      avatarId:
        type: string
      avatarUrl:
        type: string
      displayName:
        type: string
      id:
        type: string
      isOwner:
        type: boolean
      online:
        type: boolean
    type: object
  server.ServerInfo:
    properties:
      id:
        type: string
      name:
        type: string
      ownerId:
        type: string
    type: object
info:
  contact: {}
  description: Armonic API backend server.
  title: Armonic API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user in already claimed server
      parameters:
      - description: User credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: invalid body
          schema:
            type: string
        "401":
          description: invalid credentials
          schema:
            type: string
        "403":
          description: server not claimed yet
          schema:
            type: string
        "409":
          description: proof of work expired or already used
          schema:
            type: string
        "429":
          description: rate limit exceeded
          schema:
            type: string
      summary: Login user
      tags:
      - Users
...
  /server:
    get:
      description: List the servers the authenticated user is a member of
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.ServerInfo'
            type: array
        "401":
          description: unauthorized
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: My servers
      tags:
      - Server
...
  /server/{id}/invite:
    post:
      description: Owner-only. Mint a single-use invite link for a server.
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.InviteCreatedResponse'
        "401":
          description: unauthorized
          schema:
            type: string
        "403":
          description: forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create invite
      tags:
      - Invite
  /server/{id}/members:
    get:
      description: 'Member-only. The roster of a server: every member with their display
        name, avatar and owner flag, plus whether they currently have a live WebSocket
        connection.'
      parameters:
      - description: Server ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/server.MemberInfo'
            type: array
        "400":
          description: invalid server id
          schema:
            type: string
        "403":
          description: forbidden
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Server members
      tags:
      - Server
...
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and the JWT returned by /auth/login.
      Example: "Bearer eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...